solidify export_stl and use shape handles

+++ test/test.js
@@ -1,13 +1,18 @@
 var _test = require('tape');
 var fs = require('fs');
 var path = require('path');
+var pjoin = path.join;
 var protobuf = require('protocol-buffers/require');
 var spawn = require('child_process').spawn;
 var oce = protobuf('../protocol/oce.proto');
 var request = oce.NetOCE_Request;
 var response = oce.NetOCE_Response;
+var tdir = __dirname;
+var tmpdir = pjoin(tdir, 'tmp') + '/';
+var exe = pjoin(tdir, '..', 'out/bin/net-oce');
 
-var exe = path.join(__dirname, '..', 'out/bin/net-oce');
+var schema = require('../schema');
+var ENUM = schema.NetOCE_Value.type;
 
 function setup(fn) {
   var child = spawn(exe, [], {
@@ -34,23 +39,29 @@ function setup(fn) {
     r.value.forEach(function(op) {
       var id = op.operation.id;
 
-      methods[op.operation.name] = function() {
+      methods[op.operation.name] = function(a, fn) {
+        var args;
 
-        var args = [];
-        Array.prototype.push.apply(args, arguments);
-        var fn = args.pop();
-
-        var obj = {
-          method : id,
-          seq: 123123,
-          argument: args.map(function(arg) {
+        if (Array.isArray(a)) {
+          args = a;
+        } else {
+          args = [];
+          Array.prototype.push.apply(args, arguments);
+          fn = args.pop();
+          args = args.map(function(arg) {
             // TODO: do proper typing
             // TODO: Handle wrapper
             return {
               type: 1,
               double_value: arg
             }
-          })
+          });
+        }
+
+        var obj = {
+          method : id,
+          seq: 123123,
+          argument: args
         };
         queue.push(fn);
 
@@ -67,21 +78,52 @@ function setup(fn) {
 
 function test(name, fn) {
   setup(function(e, child) {
-    _test(function(t) {
+    _test(name, function(t) {
       var end = t.end.bind(t);
+      var ended = false;
       t.end = function() {
-        child._process.kill();
+        ended = true;
+        child._process && child._process.kill();
         end();
       };
+
+      child._process.on('close', function() {
+        if (!ended) {
+          t.fail("net-oce crashed");
+          t.end();
+        }
+      });
 
       fn(child, t);
     });
   });
 }
 
+test('stl export - no args', function(child, t) {
+  child.export_stl(function(e, r) {
+    t.equal(r.value[0].type, ENUM('ERROR'));
+    t.end();
+  });
+});
+
+test('stl export - invalid SHAPE_HANDLE', function(child, t) {
+  child.cube(0, 0, 0, 10, 10, 10, function(e, cube) {
+    child.export_stl([
+      { type : ENUM('SHAPE_HANDLE'), uint32_value: 0 },
+      { type : ENUM('STRING'), string_value: 'blah.stl' }
+    ], function(e, r) {
+      t.equal(r.value[0].type, ENUM('ERROR'));
+      t.end();
+    });
+  });
+});
+
 test('stl export - no shapes', function(child, t) {
-  child.export_stl(function(e, result) {
-    t.equal(result.value[0].type, 13);
+  child.export_stl([
+    { type : ENUM('SHAPE_HANDLE'), uint32_value: 123 },
+    { type : ENUM('STRING'), string_value: 'blah.stl' }
+  ], function(e, result) {
+    t.equal(result.value[0].type, ENUM('BOOL'));
     t.equal(result.value[0].bool_value, false);
 
     t.end();
@@ -91,11 +133,13 @@ test('stl export - no shapes', function(child, t) {
 test('stl export - cube', function(child, t) {
   child.cube(0, 0, 0, 10, 10, 10, function(e, cube) {
 
-    t.equal(cube.value[0].type, 18);
-    t.equal()
+    t.equal(cube.value[0].type, ENUM('SHAPE_HANDLE'));
 
-    child.export_stl(function(e, result) {
-      t.equal(result.value[0].type, 13);
+    child.export_stl([
+      { type : ENUM('SHAPE_HANDLE'), uint32_value: cube.value[0].uint32_value },
+      { type : ENUM('STRING'), string_value: tmpdir + 'blah.stl' }
+    ], function(e, result) {
+      t.equal(result.value[0].type, ENUM('BOOL'));
       t.equal(result.value[0].bool_value, true);
 
       t.end();
@@ -105,14 +149,14 @@ test('stl export - cube', function(child, t) {
 
 test('cube - no args', function(child, t) {
   child.cube(function(e, result) {
-    t.equal(result.value[0].type, 17);
+    t.equal(result.value[0].type, ENUM('ERROR'));
     t.end();
   });
 });
 
 test('cube', function(child, t) {
   child.cube(0, 0, 0, 10, 10, 10, function(e, cube) {
-    t.equal(cube.value[0].type, 18);
+    t.equal(cube.value[0].type, ENUM('SHAPE_HANDLE'));
     t.ok(cube.value[0].uint32_value !== 0);
     t.end();
   });
